feat : book의 rating double -> int 바꿈 ㅋ

+++ src/main/java/com/book4w/book4w/dto/response/HomeRecommendedResponseDTO.java
@@ -15,7 +15,7 @@ public class HomeRecommendedResponseDTO {
     private String coverImage;
 
 
-    private double bookRating;
+    private int bookRating;
     private int reviewCount;
     private int likeCount;
